Updated to maven
With great regret we have moved to maven.  The gradle support seemed to cause hurdles each time we needed to do anything, it felt like it was going against the gradle grain.  Maven seems to have better support at the moment, so we have restructured into maven modules

May the force be with us!